feat: tweaked cocaine and moved movement modifier to own file
