+++ Assets/Scripts/Larvae/Drugs/MovementModifier.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 87055914c4ed4ee6bc595e3376c3327e
+timeCreated: 1756494449
